+++ CHANGES.md
@@ -23,6 +23,7 @@
 * remove deprecated override dot org constant
 * added setting to display `_deprecated_hook()` data in debug.log
 * skip `_deprecated_hook()` `trigger_error()` in development environment
+* add upsell notice
 
 #### 9.9.10 / 2021-02-18
 * fix change to `redirect_on_save()`